Lab7-RAG/Create-LLama-Example/frontend/requirements.txt
streamlit>=1.31.0
requests>=2.31.0
python-dotenv>=1.0.0
watchdog>=3.0.0  # For better reloading experience
